+++ tests/html-internal-link-parser.test.js
@@ -7,7 +7,7 @@ const pageDirectory = pageLookup([]);
 
 test('html link parser grabs multiple href, ignoring external links', t => {
     const parser = new HTMLLinkParser(new DeadLinks());
-    const links = parser.find('<p>Hello world <a href="/home">this is a link home</a> and <a href="/somewhere">this is a link somewhere</a></p><p>The following link should be ignored <a href="http://www.example.com/">example.com</a>.</p>', pageDirectory);
+    const links = parser.find('<p>Hello world <a href="/home">this is a link home</a> and <a href="/somewhere">this is a link somewhere</a></p><p>The following link should be ignored <a href="https://www.example.com/">example.com</a>.</p>', pageDirectory);
 
     t.is(2, links.length);